requirements.txt
flask
gunicorn
pandas
numpy
matplotlib
requests
python-dotenv
openpyxl
pillow
